+++ src/ui/operation-modes/modes/object-inspection/component-inspectors/GOCTransform.cpp
@@ -1,7 +1,7 @@
 #include "GOCTransform.h"
 #include <ui/common/editors/Basic.h>
 #include <ui/common/viewers/Basic.h>
-
+#include <utilities/GameObjectUtils.h>
 
 using namespace hh::game;
 
@@ -9,15 +9,16 @@ void RenderComponentInspector(GOCTransform& component) {
 	auto transform = component.GetTransform();
 
 	ImGui::SeparatorText("Main local transform (editable form)");
-	if (Editor("GOCTransformTransformEditable", transform))
+	if (Editor("GOCTransformTransformEditable", transform)) {
 		component.SetLocalTransform(transform);
+		EnsureGOCTransformIsUpdating(component);
+	}
+
+	CheckboxFlags("Transform position", component.GetFrame().flags, hh::fnd::HFrame::Flag::TRANSFORM_POSITION);
+	CheckboxFlags("Transform rotation", component.GetFrame().flags, hh::fnd::HFrame::Flag::TRANSFORM_ROTATION);
 
 	ImGui::SeparatorText("Main local transform (internal representation)");
 	Viewer("GOCTransformTransform", component.GetTransform());
 
-	ImGui::SeparatorText("HFrame local transform");
-	Viewer("HFrameTransformLocal", component.GetFrame().localTransform);
-
-	ImGui::SeparatorText("HFrame full transform");
-	Viewer("HFrameTransformFull", component.GetFrame().fullTransform);
+	Viewer("HFrame", component.GetFrame());
 }